modify geometry shader triangle index handling

+++ Engine/DefaultGeometryShader.h
@@ -8,7 +8,7 @@ class DefaultGeometryShader
 public:
 	typedef Vertex Output;
 public:
-	Triangle<Output> operator()( const Vertex& in0,const Vertex& in1,const Vertex& in2,unsigned int triangle_index ) const
+	Triangle<Output> operator()( const Vertex& in0,const Vertex& in1,const Vertex& in2,size_t triangle_index ) const
 	{
 		return{ in0,in1,in2 };
 	}
