+++ Engine/SolidGeometryEffect.h
@@ -126,7 +126,7 @@ public:
 			Color color;
 		};
 	public:
-		Triangle<Output> operator()( const VertexShader::Output& in0,const VertexShader::Output& in1,const VertexShader::Output& in2,unsigned int triangle_index ) const
+		Triangle<Output> operator()( const VertexShader::Output& in0,const VertexShader::Output& in1,const VertexShader::Output& in2,size_t triangle_index ) const
 		{
 			return{ 
 				{ in0.pos,triangle_colors[triangle_index/2] },
